+++ oaebu_workflows/fixtures/oapen_metadata/test_table.json
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:03f7bfc4677e7cdc21e36f66dcb5f2959b568bd504bb5ecf4dba80360493a721
-size 24805
+oid sha256:e6022995307477990e992f319704d82c1a29351ad795035dd3e9ce5955157f81
+size 25944
