+++ oaebu_workflows/fixtures/onix_utils/input_metadata.xml
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8e9684060f2df66b808df90544ed52a2ab4fbf1474dee5dff97949dcdc81201b
+size 28788
